move flag base definition to its own file
