+++ src/00_Core/Item/ItemManager.cpp
@@ -9,6 +9,7 @@
 #include "Player/PlayerManager.hpp"
 #include "Save/AdventureFlags.hpp"
 #include "System/OverlayManager.hpp"
+#include "flags.h"
 
 static const char *sShipPartTypes[] = {"anc", "bow", "hul", "can", "dco", "pdl", "fnl", "brg"};
 
@@ -407,7 +408,7 @@ THUMB void ItemManager::SetTreasureSalvaged(u32 index) {
 }
 
 THUMB void ItemManager::RemoveItem(ItemFlag item) {
-    RESET_FLAG(mItemFlags.flags, item);
+    UNSET_FLAG(mItemFlags.flags, item);
     if (item >= ItemFlag_EQUIP_START && item <= ItemFlag_EQUIP_END) {
         (*mAmmo)[item] = 0;
     }
@@ -992,7 +993,7 @@ ARM void ItemManager::func_ov00_020ae4dc(s32 param1) {
 ARM void ItemManager::SetPotion(u32 index, Potion potion) {
     mPotions[index] = potion;
     if (potion == Potion_None) {
-        RESET_FLAG(mItemFlags.flags, index + ItemFlag_PotionA);
+        UNSET_FLAG(mItemFlags.flags, index + ItemFlag_PotionA);
     } else {
         SET_FLAG(mItemFlags.flags, index + ItemFlag_PotionA);
     }
